refactor: extract `openConflictPicker` from `index.ts`
