+++ __tests__/utils/conflicts-handler.spec.ts
@@ -11,7 +11,7 @@ const validateFilename = vi.hoisted(() => vi.fn(() => true))
 const openConflictPicker = vi.hoisted(() => vi.fn())
 const showInvalidFilenameDialog = vi.hoisted(() => vi.fn())
 
-vi.mock('../../lib/index.ts', () => ({ openConflictPicker }))
+vi.mock('../../lib/ui/openConflictPicker.ts', () => ({ openConflictPicker }))
 vi.mock('../../lib/ui/utils/dialog.ts', () => ({ showInvalidFilenameDialog }))
 vi.mock('@nextcloud/files', async (getModule) => {
 	const original = await getModule()
